requirements.txt
streamlit==1.21.0
pyrebase4==4.5.0
requests==2.28.2